feat: user can persist task description changes from single task expanded view, next add and persist change description, activity badge
