+++ client/context/projectContext.js
@@ -1,4 +1,5 @@
 import React, {useState, useEffect, useMemo} from 'react'
+import {getColumnsDB} from './axiosService'
 import axios from 'axios'
 
 export const ProjectContext = React.createContext()
@@ -32,6 +33,23 @@ export default function ProjectProvider({projectId, children}) {
     }
   }, [taskChanged])
 
+  const refreshProjectBoard = async () => {
+    // important! we set columns on project context
+    // so that local task state persists across
+    // drag and drop -- to do that, we need to
+    // first store the current column ORDER
+    // and rearrange our fetched columns from the db
+    const currentColumnOrder = columns.map((col) => col.id)
+
+    const fetchedColumns = await getColumnsDB(project.id)
+
+    const updatedColumns = currentColumnOrder.map((orderedId) =>
+      fetchedColumns.find((col) => col.id === orderedId)
+    )
+
+    setColumns(updatedColumns)
+  }
+
   const providerValue = useMemo(() => {
     return {
       project,
@@ -44,6 +62,7 @@ export default function ProjectProvider({projectId, children}) {
       setComments,
       taskChanged,
       setTaskChanged,
+      refreshProjectBoard,
     }
   }, [project, columns, tasks, comments])
